requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
upstash-redis
python-dotenv
